cart page and all the functionalities are added
