+++ cart.html
@@ -4,15 +4,16 @@
     <meta charset="UTF-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Document</title>
+    <title>My cart</title>
     <link rel="preconnect" href="https://fonts.googleapis.com">
       <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
       <link href="https://fonts.googleapis.com/css2?family=Lato:wght@300;400&display=swap" rel="stylesheet">
       <link rel ="stylesheet" href = "style.css">
-      <script type="module" src = "index.js" defer></script>
+     <script type="module" src = "scripts/cart.js" defer></script>
 </head>
 <body>
-    <!-- <div id = "added-items">
-    </div> -->
+    <h1>Cart</h1>
+    <div id = "added-items-cart">
+    </div>
 </body>
 </html>
