+++ scripts/data_stub.js
